+++ src/backend/resolvers/mutations/collection/uploadSheet/importPayerHistoricalData/index.js
@@ -1,3 +1,0 @@
-const importPayerHistoricalData = require('./importPayerHistoricalData')
-
-module.exports = importPayerHistoricalData