+++ src/backend/resolvers/mutations/collection/uploadSheet/importPayerHistoricalLivesData/index.js
@@ -1,0 +1,3 @@
+const importPayerHistoricalLivesData = require('./importPayerHistoricalLivesData')
+
+module.exports = importPayerHistoricalLivesData